# customer-service/src/main/resources/openapi/customer-api.yaml
openapi: 3.0.3
info:
  title: Customer Service API
  description: API for managing customers and persons
  version: 1.0.0
  contact:
    name: NTT Data
    email: support@nttdata.com

servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Customers
    description: Customer management operations

paths:
  /api/v1/customers:
    get:
      tags:
        - Customers
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
    post:
      tags:
        - Customers
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{id}:
    get:
      tags:
        - Customers
      summary: Get customer by ID
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Customers
      summary: Update customer
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Customers
      summary: Delete customer
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CustomerRequest:
      type: object
      required:
        - name
        - gender
        - identification
        - address
        - phone
        - password
        - status
      properties:
        name:
          type: string
          example: "Jose Lema"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          example: "MALE"
        identification:
          type: string
          example: "1234567890"
        address:
          type: string
          example: "Otavalo sn y principal"
        phone:
          type: string
          example: "098254785"
        password:
          type: string
          format: password
          example: "1234"
        status:
          type: boolean
          example: true

    CustomerResponse:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        name:
          type: string
        gender:
          type: string
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        status:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string